Stop player when extensions are added (#448)
**User-Facing Changes**
With these changes when a user tries to add extensions when the player
is running, the player is stopped and an overlay appears informing the
user that extensions are loading. After that process is done the overlay
disappears again and the player resumes.


![image](https://github.com/user-attachments/assets/e2bc5756-f9b0-4c7b-a28e-7b7ca399219e)



**Description**
Refactored existing logic into a new hook `useHandleFiles`, and
introduced a new hook `useInstallingExtensionsState` to manage the
extension installation process in Lichtblick.

Additionally, created a dedicated store to track installation progress,
including a new `inProgress` flag that indicates whether an extension is
currently being installed. This flag is now used in `PanelLayout` to
conditionally render the overlay shown during installation (as seen in
the screenshot above).

<!-- link relevant GitHub issues -->
<!-- add `docs` label if this PR requires documentation updates -->
<!-- add relevant metric tracking for experimental / new features -->

**Checklist**

- [X] The web version was tested and it is running ok
- [X] The desktop version was tested and it is running ok
- [X] This change is covered by unit tests
- [X] Files constants.ts, types.ts and *.style.ts have been checked and
relevant code snippets have been relocated